feat(cli): move 'helm dm' to 'helm server'
