--- cmd/helm/target.go
@@ -1,37 +1,0 @@
-/*
-Copyright 2016 The Kubernetes Authors All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package main
-
-import (
-	"fmt"
-
-	"github.com/kubernetes/helm/pkg/format"
-	"github.com/kubernetes/helm/pkg/kubectl"
-)
-
-func target(dryRun bool) error {
-	client := kubectl.Client
-	if dryRun {
-		client = kubectl.PrintRunner{}
-	}
-	out, err := client.ClusterInfo()
-	if err != nil {
-		return fmt.Errorf("%s (%s)", out, err)
-	}
-	format.Msg(string(out))
-	return nil
-}
